tts-azure/requirements.txt
fastapi
pyyaml
aiohttp
sox
watchdog
argh
